+++ usi-wordpress-solutions-history.php
@@ -15,9 +15,11 @@ https://github.com/jaschwanda/wordpress-solutions/blob/master/LICENSE.md
 Copyright (c) 2020 by Jim Schwanda.
 */
 
+require_once('usi-wordpress-solutions.php');
+
 final class USI_WordPress_Solutions_History {
 
-   const VERSION = '2.9.5 (2020-09-14)';
+   const VERSION = '2.9.10 (2020-10-16)';
 
    private static $pre_post_update_data = null;
    private static $pre_post_update_id   = 0;
@@ -30,15 +32,19 @@ final class USI_WordPress_Solutions_History {
 
    public static function _init() {
 
-      add_action('delete_post', array(__CLASS__, 'action_delete_post'));
-      add_action('delete_user', array(__CLASS__, 'action_delete_user'), 10, 2);
-      add_action('edit_user_profile_update', array(__CLASS__, 'action_profile_update'));
-      add_action('pre_post_update', array(__CLASS__, 'action_pre_post_update'), 110, 2);
-      add_action('user_register', array(__CLASS__, 'action_user_register'), 10, 2);
-      add_action('wp_insert_post', array(__CLASS__, 'action_wp_insert_post'), 10, 3);
-      add_action('wp_login', array(__CLASS__, 'action_wp_login'), 10, 3);
+      if (!empty(USI_WordPress_Solutions::$options['admin-options']['history'])) {
 
-      add_filter('logout_redirect', array(__CLASS__, 'filter_logout_redirect'), 10, 3);
+         add_action('delete_post', array(__CLASS__, 'action_delete_post'));
+         add_action('delete_user', array(__CLASS__, 'action_delete_user'), 10, 2);
+         add_action('edit_user_profile_update', array(__CLASS__, 'action_profile_update'));
+         add_action('pre_post_update', array(__CLASS__, 'action_pre_post_update'), 110, 2);
+         add_action('user_register', array(__CLASS__, 'action_user_register'), 10, 2);
+         add_action('wp_insert_post', array(__CLASS__, 'action_wp_insert_post'), 10, 3);
+         add_action('wp_login', array(__CLASS__, 'action_wp_login'), 10, 3);
+
+         add_filter('logout_redirect', array(__CLASS__, 'filter_logout_redirect'), 10, 3);
+
+      }
 
    } // _init();
 
@@ -137,14 +143,20 @@ final class USI_WordPress_Solutions_History {
    } // from();
 
    public static function history($user_id, $type, $action, $target_id = 0, $data = null) {
-      global $wpdb;
-      if (is_array($data) || is_object($data)) $data = substr(print_r($data, true), 0, 65535);
-      if (false === $wpdb->insert(
-         $wpdb->prefix . 'USI_history', 
-         array('user_id' => $user_id, 'type' => $type, 'action' => $action, 'target_id' => $target_id, 'data' => $data),
-         array('%d', '%s', '%s', '%d', '%s'))) {
-         usi::log2('last-error=', $wpdb->last_error);
+
+      if (!empty(USI_WordPress_Solutions::$options['admin-options']['history'])) {
+
+         global $wpdb;
+         if (is_array($data) || is_object($data)) $data = substr(print_r($data, true), 0, 65535);
+         if (false === $wpdb->insert(
+            $wpdb->prefix . 'USI_history', 
+            array('user_id' => $user_id, 'type' => $type, 'action' => $action, 'target_id' => $target_id, 'data' => $data),
+            array('%d', '%s', '%s', '%d', '%s'))) {
+            usi::log2('last-error=', $wpdb->last_error);
+         }
+
       }
+
    } // history();
 
 } // Class USI_WordPress_Solutions_History;